meta: The great re-org, PT 1
