+++ .gitignore
@@ -91,12 +91,6 @@ Temporary Items
 # Created by https://www.toptal.com/developers/gitignore/api/rust
 # Edit at https://www.toptal.com/developers/gitignore?templates=rust
 
-### Rust ###
-# Generated by Cargo
-# will have compiled files and executables
-debug/
-target/
-
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
 Cargo.lock
