+++ src/hal/include/hal/hal_logger.h
@@ -1,0 +1,7 @@
+#pragma once
+
+#include <logger.h>
+
+extern logger_t hal_logger;
+
+void hal_logger_initialize();
