+++ src/hal/include/hal/serial.h
