+++ src/kernel/xmake.lua
@@ -1,26 +1,28 @@
 target("kernel")
     set_kind("binary")
-    set_arch("x86_64")
     add_files("core/*.c")
-    add_files("core/*.asm", {force = {nasm = true}})
-    add_values("nasm.flags", "-felf64")
     set_filename("kernel")
 
-    add_includedirs("$(projectdir)/src/kernel/include")
+    add_includedirs("$(projectdir)/src/hal/include")
+    add_includedirs("$(projectdir)/src/libs/libk/include")
     add_includedirs("$(projectdir)/src/libs/limine/include")
     add_includedirs("$(projectdir)/src/libs/ssfn/include")
     add_includedirs("$(projectdir)/src/libs/logger/include")
     add_includedirs("$(projectdir)/src/libs/printf")
+    add_includedirs("$(projectdir)/src/kernel/include")
 
     add_ldflags("-T$(projectdir)/meta/linker.ld", { force = true })
-    add_ldflags("-pie", { force = true })
     add_ldflags("-z text", { force = true })
     add_ldflags("-z max-page-size=0x1000", { force = true })
 
+    add_deps("libk")
+    add_deps("printf")
+    add_deps("hal")
+    add_deps("logger")
     add_deps("limine")
     add_deps("ssfn")
-    add_deps("logger")
-    add_deps("printf")
+
+    --add_links("libk", "printf", "logger", "hal")
 
     -- Debug configuration (always on for now)
     add_cxflags("-g")
