+++ src/libs/libk/include/libk/string.h
