+++ src/libs/libk/src/string.c
@@ -2,7 +2,7 @@
 #include <stddef.h>
 #include <stdbool.h>
 
-#include <kernel/string.h>
+#include <libk/string.h>
 
 void *memcpy(void *dest, const void *src, size_t n) {
   uint8_t *pdest = (uint8_t *)dest;
