+++ src/libs/libk/xmake.lua
@@ -1,0 +1,5 @@
+target("libk")
+    set_kind("static")
+    add_files("src/*.c")
+    add_includedirs("include", {public = true})
+    add_defines("LIBK_BUILD", {public = true})
